Made rank space breeding incorporate the interval tree.

+++ genetic_algorithm/src/evaluation/Specimen.java
@@ -6,13 +6,13 @@ public interface Specimen {
      * 
      * @return the score for the current specimen
      */
-    public float getEvaluatedScore();
+    public double getEvaluatedScore();
     
     /**
      * 
      * @param s
      * @return the distance between specimen s and the current specimen
      */
-    public float getDistance(Specimen s);
+    public double getDistance(Specimen s);
     
 }
